requirements.txt
Flask==3.0.2
openai==0.28.1
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
Flask-Limiter[redis]==3.5.0
redis==5.0.1 